+++ relay-server/src/metrics_extraction/sessions.rs
@@ -4,9 +4,44 @@ use relay_common::{UnixTimestamp, Uuid};
 use relay_general::protocol::{
     AbnormalMechanism, SessionAttributes, SessionErrored, SessionLike, SessionStatus,
 };
-use relay_metrics::{Metric, MetricValue, SessionsKind};
+use relay_metrics::{Metric, MetricValue, SessionsKind, TypedMRI};
 
 use super::utils::with_tag;
+
+/// Enumerates the most common session-names.
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum SessionsKind {
+    Session,
+    User,
+    Error,
+}
+
+impl SessionsKind {
+    pub fn ty(&self) -> MetricType {
+        match self {
+            Self::Session => MetricType::Counter,
+            Self::Error => MetricType::Set,
+            Self::User => MetricType::Set,
+        }
+    }
+
+    pub fn unit(&self) -> MetricUnit {
+        MetricUnit::None
+    }
+    pub fn to_mri_string(self) -> String {
+        TypedMRI::from(self).to_string()
+    }
+}
+
+impl Display for SessionsKind {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Session => write!(f, "session"),
+            Self::User => write!(f, "user"),
+            Self::Error => write!(f, "error"),
+        }
+    }
+}
 
 /// Convert contained nil UUIDs to None
 fn nil_to_none(distinct_id: Option<&String>) -> Option<&String> {
@@ -47,8 +82,8 @@ pub fn extract_session_metrics<T: SessionLike>(
     // Always capture with "init" tag for the first session update of a session. This is used
     // for adoption and as baseline for crash rates.
     if session.total_count() > 0 {
-        target.push(Metric::new_mri(
-            SessionsKind::Session.into(),
+        target.push(Metric::new(
+            TypedMRI::from(SessionsKind::Session).to_string(),
             MetricValue::Counter(session.total_count() as f64),
             timestamp,
             with_tag(&tags, "session.status", "init"),
@@ -58,14 +93,14 @@ pub fn extract_session_metrics<T: SessionLike>(
     // Mark the session as errored, which includes fatal sessions.
     if let Some(errors) = session.all_errors() {
         target.push(match errors {
-            SessionErrored::Individual(session_id) => Metric::new_mri(
-                SessionsKind::Error.into(),
+            SessionErrored::Individual(session_id) => Metric::new(
+                SessionsKind::Error.to_mri_string(),
                 MetricValue::set_from_display(session_id),
                 timestamp,
                 tags.clone(),
             ),
-            SessionErrored::Aggregated(count) => Metric::new_mri(
-                SessionsKind::Session.into(),
+            SessionErrored::Aggregated(count) => Metric::new(
+                SessionsKind::Session.to_mri_string(),
                 MetricValue::Counter(count as f64),
                 timestamp,
                 with_tag(&tags, "session.status", "errored_preaggr"),
@@ -73,8 +108,8 @@ pub fn extract_session_metrics<T: SessionLike>(
         });
 
         if let Some(distinct_id) = nil_to_none(session.distinct_id()) {
-            target.push(Metric::new_mri(
-                SessionsKind::User.into(),
+            target.push(Metric::new(
+                SessionsKind::User.to_mri_string(),
                 MetricValue::set_from_str(distinct_id),
                 timestamp,
                 with_tag(&tags, "session.status", "errored"),
@@ -84,8 +119,8 @@ pub fn extract_session_metrics<T: SessionLike>(
         // For session updates without errors, we collect the user without a session.status tag.
         // To get the number of healthy users (i.e. users without a single errored session), query
         // |users| - |users{session.status:errored}|
-        target.push(Metric::new_mri(
-            SessionsKind::User.into(),
+        target.push(Metric::new(
+            SessionsKind::User.to_mri_string(),
             MetricValue::set_from_str(distinct_id),
             timestamp,
             tags.clone(),
@@ -95,8 +130,8 @@ pub fn extract_session_metrics<T: SessionLike>(
     // Record fatal sessions for crash rate computation. This is a strict subset of errored
     // sessions above.
     if session.abnormal_count() > 0 {
-        target.push(Metric::new_mri(
-            SessionsKind::Session.into(),
+        target.push(Metric::new(
+            SessionsKind::Session.to_mri_string(),
             MetricValue::Counter(session.abnormal_count() as f64),
             timestamp,
             with_tag(&tags, "session.status", SessionStatus::Abnormal),
@@ -112,8 +147,8 @@ pub fn extract_session_metrics<T: SessionLike>(
                     session.abnormal_mechanism().to_string(),
                 );
             }
-            target.push(Metric::new_mri(
-                SessionsKind::User.into(),
+            target.push(Metric::new(
+                SessionsKind::User.to_mri_string(),
                 MetricValue::set_from_str(distinct_id),
                 timestamp,
                 tags_for_abnormal_session,
@@ -122,16 +157,16 @@ pub fn extract_session_metrics<T: SessionLike>(
     }
 
     if session.crashed_count() > 0 {
-        target.push(Metric::new_mri(
-            SessionsKind::Session.into(),
+        target.push(Metric::new(
+            SessionsKind::Session.to_mri_string(),
             MetricValue::Counter(session.crashed_count() as f64),
             timestamp,
             with_tag(&tags, "session.status", SessionStatus::Crashed),
         ));
 
         if let Some(distinct_id) = nil_to_none(session.distinct_id()) {
-            target.push(Metric::new_mri(
-                SessionsKind::User.into(),
+            target.push(Metric::new(
+                SessionsKind::User.to_mri_string(),
                 MetricValue::set_from_str(distinct_id),
                 timestamp,
                 with_tag(&tags, "session.status", SessionStatus::Crashed),
